(not-working) text, and top/right buttons
Tempted to do a function that takes numbers and letters, but a lot of
launchpad owners will have covered those up with ableton stickers, so
maybe not.

I should now validate inputs, either ignore out-of-bounds or throw.

Then, figure out the approach for react-like modeling of the whole
board. While still allowing for realtime one-at-a-time updates. Also
would be nice to have the launchpad handle tied up in that state.