Release 0.1.0
Tweak the docs, README, etc. This is a development release but is enough
to do blinkenlights. I may change some names (Model in particular, to
I-don't-know-what; and maybe `side` to `right`) but the idea should be
pretty stable.

Next up: figure out responding to events.

I'm a clojure n00b so be forgiving of any non-idiomatic code. Even
better, tell me how you'd do it!