a sketch of a spectrum analyzer example
this doesn't actually interact with the pad yet, I'm just focused on
the audio capture and calculation part. As-is it seems to do the
calculations but takes almost a minute and a half to do the 4096-point
mdct. So yeah, that ain't working. I guess I'll draft up a basic fft and
hope that's plenty fast, otherwise I might be doing something like jni
or zmq ipc to a program in a fast language.