Let this be a lesson to you, my son
I should have just started with the FFT. It was not as hard to code as I
thought - no harder than the mDCT really. (This is not an optimized FFT,
just the basic butterfly radix-2 algorithm from wikipedia)

It uses a silly amount of CPU for what it's doing, but has no trouble
keeping up. I was confused for awhile about getting twice the framerate
I expected until I realized I was getting 16-bit samples. I wasn't sure
how to go from a byte array to shorts with clojure so I just asked for
8-bit samples. (I did try ByteBuffer to ShortBuffer and boy howdy did
that slow things down ridiculously. I may try manually converting and
see if that's faster - it should be)