begin again on tests
A test! Most of the work here was getting the newly refactored stuff to
compile. Found a bug in the implementation of decode-midi for :side
though. Yay tests!

Moved the stuff I don't want to focus on away, out of the sight of
`lein quickie`, but it'll probably move back later.