Update randgrid and docs
Since Max found the problem with perf, there's not a huge difference
between a sequence of updates and updating a whole state change at once.
Amazing what an 80x speedup will do. ;-)

So I updated randgrid to be unrelated to perf and just a fun little
random blinkenlights program. And updated the README.